Refactoring: Move task creation to KotlinFactories
so as to prepare for migrating to KGP API
